metamask/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
bitarray==2.8.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cytoolz==0.12.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
Faker==19.12.0
Flask==3.0.0
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
MarkupSafe==2.1.3
mnemonic==0.20
multidict==6.0.4
parsimonious==0.9.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.4
pycryptodome==3.19.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
stripe==7.1.0
toolz==0.12.0
typing_extensions==4.8.0
urllib3==2.0.7
web3==6.11.2
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
